requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mangum==0.17.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
google-generativeai==0.3.2
reportlab==4.0.7
boto3==1.34.0
